Add containerapp tests
